Rebuild the package-lock.json
Some developers have started to get npm install errors on windows, which
seems to be related to incomplete optional dependencies in the package-
lock.json file. This is a known issue that is linked to on nx's
toubleshooting page: https://github.com/npm/cli/issues/4828

The solution is to delete the node_modules folder and the package-lock
file, and then run npm install again to rebuild it.

Per Inge Vaaje <60141320+PerIngeVaaje@users.noreply.github.com>
Rune Johansen <>