Add dependabot config - dependency version control (#48)
Since we want to have Dependabot help with updating versions of our
dependencies, we need to add the configuration file for this.

However, we also have to make sure the dependencies are not updated past
what is supported by Nx. Therefore we have opted to use the
lockfile-only update strategy, and do manual updates also.