Feature/pxweb2-41 install i18next (#61)
* Added translation files and i18n config

Since we want to support multiple languages in the application, we need
to add translation files for the languages we want to support. We also
need to add a configuration file for i18next, which is the library we
use for translations.

* Remove caching of translation files

Since site administrators may want to change translations we cannot
cache the translation files.

* Add direction and arabic language file

Since we need to support right-to-left languages, we need to implement
support for document direction. We also added arabic as a language,
since we need an rtl language.

We did not add plurals or numbers or date support yet.

Mikael Nordberg <>

* Add basic date handling

Since we want to support changing how dates are formatted, we need to
set that up with i18next. This commit adds only a very basic date
formatter, but it might be enough.

Mikael Nordberg <>

* Add a date format with hours/minutes

Add a date formatted with both the date and the time of day.

Mikael Nordberg <>

* Add a suspense to the main app, to show loading

Add a basic suspense component around the App, to show loading state for
requests the server makes.

Co-authored-by: Mikael Nordberg <>

* Remove date formatting from non-default translations

Since the date formatting works the same in all languages, we do not
need separate date formatting for different languages. i18next uses the
right one even if it is only specified in the fallback language.

---------

Co-authored-by: MikaelNordberg <mikael.nordberg@scb.se>