Add custom number formatting to examples
Add an example of how to override the formatting options, so we can have
other decimal numbers or rounding modes than what is specified in the
translation.json file.

This also gives us the ability to set the options based on a variable.
The next step is to look into turning this into a component to be
reused, so we can have global settings for the options.

Co-authored-by: Mikael Nordberg <>