Add custom formatting of numbers
Since we want to format numbers based on the language chosen, we need to
add this to the project. The basic functionality was added when i18next
was installed, but we have added examples for how to use this in a few
different ways.

We noticed that a basic example led to many places to set the
roundingMode, and therefore created a function and a cusstom formatter
for i18next, where we set this roundingMode based on a variable. This
prepares the number formatting for the work to be done later, with a
config file for the project.

Co-authored-by: Mikael Nordberg <>