Add breakpoints tokens
Since we need to control the layout for different sized screens and
such, we need the breakpoints for the media queries. This commit adds
the breakpoints to a file, and adds that to the style-dictionary
build command. The same as how we define and use the Spacings. In this
way we have 1 place where all the breakpoint values are set, and can
change them or add to them later if needed.