Add example of breakpoints to app
Since we want to see the breakpoints working, this commit adds an
example to the main app page.

One thing to consider is that currently the breakpoint tokens are
imported (@use) from the files that are generated in the ui lib, and
not from inside the app itself. I am not sure if we want this, or to
create another set of files inside the app's own folder, and import
from there. This is very easy to add in our current implementation,
just add another output file to the build.js in the lib.