Update dependencies with nx migrate latest
Since the Nx tool and plugins are out of date, again, by a minor
version, we need to update them. Ran the 'nx migrate latest' and
'npm update --save' commands, to update the project to the newest Nx
versions. This also updates other dependencies in the package.json.

The process created a migrations.json file, but no changes were made
when I ran the migration command. Added it to the repo just in case, but
it can be deleted again soon.

Tested that the app ran as usual, and that 'npm ci' ran fine.