Fix warnings in example app
Since we don't want unnecessary warnings and such cluttering the dev
experiance in the browser, we should do easy fixes for warnings. Even
if the site in question is a temporary one. This fixes some warnings
that were clutting the console and dev tools.

PS: My first attempt at commiting a chunk of a file, instead of the
whole file. Hopefully it works.