PXWEB2-55 Update translation files
Since we want the new components to use the translation files, so we
don't have to add that later, we need to update the keys in said files.
This also adds the translations needed for the VariableBox for en, no,
sv and ar. Though the translations were gotten from google translate
where I did not know the language.