Fix hover and keyboard nav highlighting
Since we want a good experience when using the entire site with a
keyboard, we need to have good navigation and highlighting for this.

This changes the VariableBox to not use the Button component, but
instead uses the Icon component directly and styles it based on hover
and focus etc. TabIndex was also added to enable keyboard navigation to
the komponent Header, so one can activate it. It opens/closes when using
the Enter and Space keys.

The current state of showing that the focus is on the Header component
might not be enough, so will bring this up with the designer.