PxWeb2-55 Update checkbox logic to include var ID
Since we need to be able to store the state outside of the variable box,
we need to be able to know which variable has which values and codelist
chosen. This adds the last logic for the checkbox part of this logic.

Still missing logic to add selected codelist to the state of variables.

The code can also be improved, but pushing this since we need to get a
basic version working so others can use it to develop other parts of the
tabel view.