Refactor checkboxes logic to be more readable
Since we want the codebase to be as readable and reusable as possible,
I have refactored the logic for handling the checkboxes to be more
inline with this. Alot of the logic was in the handler functions
themselves, but is now in other functions that can be reused later. For
example, when we want to add the selected values of the default
algorithm.