Fix resetting state of VB child components
Since we want to clear the state completely when we load a new table
into the table view page, currently app.tsx, we need to also clear the
'isOpen' state and the internal state of the Select component in the
VariableBox component.

Setting the 'isOpen' state to false when the tableId changes fixes both
of these faults in the page.