Remove defaults usage on CSS variable use
Since we want to be sure the styling is what it should be, we should
avoid using defaults when we use CSS vars in the scss code. We want to
be sure the styling comes from the variable, and not some semi-hidden
default somewhere.

This removes the defaults from all the places they were used in the
css for the search component.