PxWeb2-209: Fix keyboard navigation
Since many of our user use a keyboard to navigate on the site, the
VariableBox component needs to handle this correctly. This adds the
navigation order according to the Figma spec. The arrow down key is used
to enter into the list of checkboxes for the variables, and moves the
focus to the first checkbox in the list. The arrow up key is then used
to leave it again, which moves the focus back to the list.

One thing that isn't handled, is when a user moves into a list with the
arrow down key, clicks out with the mouse, and then moves back to that
list with the keyboard. It will then go into that list, instead of to
the next variable. I deemed this to be something we can fix later, since
you should not ever have this problem with only keyboard navigation.