PxWeb2-209: Add aria label and description
Since we want keyboard users to under stand how to use the different
parts of the VariableBox, we need to inform them of what the list of
values is, and how to use it.

The aria-description is coming fully in WAI-ARIA 1.3, but browser
support is at 88% according to caniuse.com. Should we change this to be
in the aria-label instead? It feels too long to be appropiate there.

Or if there is a better way to tell the users about navigating into and
out of the values list, we can change it.