Refactor keyboard navigation to design spec
Since the design spec was different from the first implementation, the
keyboard navigation was not as intended. This changes the keyboard
navigation to be what is described in Figma.

We keep track of which item in the list you moved to earler, so we can
put the focus on this item later. The list items themselves have -1
tabIndex, and the keyboard navigation for them is done is a custom
function. Therefore you can tab right out.

This could also be expanded for batch selection later.

Arrow down when the values list is focused moves you into the list.
Arrow up and down move you inside the list.
Tab and Shift+Tab move you out of the list, like normal.