PxWeb2-209: Fix keyboard navigation (#180)
* PxWeb2-209: Fix keyboard navigation

Since many of our user use a keyboard to navigate on the site, the
VariableBox component needs to handle this correctly. This adds the
navigation order according to the Figma spec. The arrow down key is used
to enter into the list of checkboxes for the variables, and moves the
focus to the first checkbox in the list. The arrow up key is then used
to leave it again, which moves the focus back to the list.

One thing that isn't handled, is when a user moves into a list with the
arrow down key, clicks out with the mouse, and then moves back to that
list with the keyboard. It will then go into that list, instead of to
the next variable. I deemed this to be something we can fix later, since
you should not ever have this problem with only keyboard navigation.

* PxWeb2-209: Add aria label and description

Since we want keyboard users to under stand how to use the different
parts of the VariableBox, we need to inform them of what the list of
values is, and how to use it.

The aria-description is coming fully in WAI-ARIA 1.3, but browser
support is at 88% according to caniuse.com. Should we change this to be
in the aria-label instead? It feels too long to be appropiate there.

Or if there is a better way to tell the users about navigating into and
out of the values list, we can change it.

* Refactor keyboard navigation to design spec

Since the design spec was different from the first implementation, the
keyboard navigation was not as intended. This changes the keyboard
navigation to be what is described in Figma.

We keep track of which item in the list you moved to earler, so we can
put the focus on this item later. The list items themselves have -1
tabIndex, and the keyboard navigation for them is done is a custom
function. Therefore you can tab right out.

This could also be expanded for batch selection later.

Arrow down when the values list is focused moves you into the list.
Arrow up and down move you inside the list.
Tab and Shift+Tab move you out of the list, like normal.

* Update accessibility translations

Also added total number of values to accessibility translations for the
list of values/checkboxes in the VariableBox.