Add the alert to VariableBox for error handling
Since we want to show when there are no selected values for a mandatory
variable, we need to add this capability to the VariableBox component.

When no value is selected, it now shows and alert in the header of the
component, and the border is colored red. This is removed when a value
is chosen.

API handling is not added, that would be a different task.