Feature/px web2 231 add alert to variablebox (#182)
* Add new Nx folders to ignore files

* Add the alert to VariableBox for error handling

Since we want to show when there are no selected values for a mandatory
variable, we need to add this capability to the VariableBox component.

When no value is selected, it now shows and alert in the header of the
component, and the border is colored red. This is removed when a value
is chosen.

API handling is not added, that would be a different task.

* Add first version of error border blinking

Since we want to make it easier to see when a mandatory variable is has
no selected values, we add a blinking animation to the error border.

The designer needs to look at this before merging.

* Fix naming and structure

* Remove animation on border for error

Remove the animation on the border when the error is shown. It was clear
enough that there is an error without it, and with the animation it
became a bit too much. Might give the wrong impression that the user has
done something wrong. The designer thought.