Fix for red error showing when loading metadata
Since we do not want the users to see a red error whenever we load
metadata for mandatory variables, we need to make sure we block the
error when metadata is loading.

This needs to be looked at again, when we look into proper loading
states for the page, since this might be not needed when that is
implemented.

Also, this is missing the logic in the Alert task that has not been
merged yet. That will be added before a PR is created for this.