Fix/pxweb2-55 fix checkboxes having wrong height in var box list (#196)
There is a problem with the height of the checkboxes in the variable box's list of values, if a checkbox has more than one line of text. The problem turned out to be because the checkboxes had a min-height, while the list they were in in the variable box had a max-height.

Removing the min-height, or the max-height fixes the problem. Since the list in the variable box needs a max-height to grow up to, and start scrolling when it grows past, we can not remove that one.

Looking in Storybook and Chromatic I can not see any change being detected when removing the min-height on the checkbox. Therefore, removing that might be the solution here, unless there is a problem here I am not aware of.