Feature/pxweb2 284 Time periods sometimes in the wrong order in the table (#216)
* feat: Update and synchronize VB values with selected context

This change addresses the issue of time periods sometimes sorted in the wrong order in the table in Pxweb2 (#284).

If selectedVBValues was not in sync with variables context, a data fetch without selected variables and values (default selection) was made. In this call time periods was sorted descending.

The fix assures that selectedVBValues are always in sync with variables context and also a call to fetch data is only made when variables context is updated. In this way a call to get data from the api always contains information about selected variables and values.

The fix also adresses a problem we had when switching table. Before the selected values for table 1 was used when fetching data for table 2. This caused a http 400 error - Non-existent variable. We were trying to get values that existed in table 1 but not exists in table 2. The problem was that the useEffect was triggered by both tableId, selectedVBValues and variables. Now we have limited the useEffect to only be triggered when variables (that is in sync with selectedVBValues) or language is changed.

* Formatted document

* To force a new build

* Removed commented out lines

* Force build and publish to Chromatic

* Update README.md

---------

Co-authored-by: PerIngeVaaje <60141320+PerIngeVaaje@users.noreply.github.com>
Co-authored-by: PerIngeVaaje <per.inge.vaaje@ssb.no>