fix/clean up app.tsx (#233)
Since we are using an errorMsg state in the App.tsx, we need to actually
do something with it when it is not an empty string. Currently we are
(potentially) setting that state at several places in the file, but
never logging any errors to the console. This adds simple logging of
that state when not an empty string, like in the TableProvider, with
console.error and an indication of which file it originates from.

* Cleanup imports
* Remove unused state
* Add logging the errorMsg to console if not empty
* Remove unnecessary blank lines
* Simplify VBreset when loading metadata