Add typescript types to style dictionary build
Since we in the EmptyState need to use the breakpoints in the code, we
need to import them from the JavaScript file that the style dictionary
builds. This build step was missing the config for also building the
TypeScript types of the files. This commit adds the needed config to
enable this.

Also added the needed config to output the color information from the
build.