Add illustration support to the EmptyState
Since we need to have illustrations in the EmptyState component, we need
to expand on it to be able to handle this. This implementation is
similar to how Icon handles it's Icons, but I am not sure if this is how
we want it or not.

This is a first draft, and there is still alot of polish left to do.