Refactor illustrations
Refactoring the illustrations to contain more of the information for
each version, after discussions with other devs and the designer.

Now the versions contain the viewBox sizes, and also have colors
imported from the CSS of the theme, instead of being hardcoded. We will
evaluate at a later time, if we should rename those colors or not, since
we want users to be able to change the colors in the illustrations.