Add so VBox reverses values order when time var
Since we want the newest values shown first when a variable is of type
'time variable', we need to reverse the order of the values for that
variable, before we render it. This is possible to keep this simple,
because the API always returns the oldest values first.

Also added two tests, to check the order of the checkboxes rendered.