Update Nx deps including storybook v8 migration
Since we want to keep the project up-to-date, we need to update the Nx
project dependencies to the latest versions. Nx have now added official
support for the newest version of Storybook, v8, which we need to turn
on certain features in Chromatic. The automigration was a bit buggy, so
had to do some things manually.