Fix/refactor rendering of variablebox in app (#250)
* Refactor how the VariableBox is used

Since we want to refactor the app.tsx, one of the things we can improve
is how we use the VariableBox component. We mapped over the variables
and created the VariableBoxes directly in the place it was used. This
adds a VariableBoxList component that abstracts that away.

There is still more work to be done here, with minimizing prop-drilling.
But this does improve the VariableBox by also adding some basic testing.

* Rename the new component