Fix the global styling for the pxweb2-ui library
Since we are resetting some of the CSS in the app, we want the stories
in storybook to reflect the same reset. Currently the reset is only done
in the app, which means the components look different between the app
and the library.

This change might result in many changes in the stories in the lib, but
should not result in any changes in the app.