Feature/pxweb2-207: Add illustration support to empty state (#240)
Since we need to have illustrations in the EmptyState component, we need to expand on it to be able to handle this. This implementation is similar to how Icon handles it's Icons. The versions contain the viewBox sizes, and also have colors imported from the CSS of the theme. We will evaluate at a later time if we should rename those colors or not, since we want users to be able to change the colors in the illustrations.

In the EmptyState we need to use the breakpoints in the code, therefore we need to import them from the JavaScript file that the style dictionary builds. This build step was missing the config for also building the TypeScript types of the files. This also adds the needed config to enable this. Also added the needed config to output the color information from the build.

* Add typescript types to style dictionary build

* Add illustration support to the EmptyState

* Add stories

* Update translations

* Add missing pretest script

* Fix for rendering SVGs in EmptyState in Safari

Since Safari (webkit based browsers) will not show an SVG without a fixed height or width set, we need to set it. 'fit-content' is used, which has a 95%+ on caniuse.com.