Cleanup for showing the others
Since I want the others to look at this test of the react router for
loading the inital data, I need to clean up the branch a bit.

There are still problems to solve, but the basic loading of inital data
in parallell should now work. It also works with making selections for
new api calls, once the inital data has finished loading.

I did not spend much time trying to get Suspense to show, or styling
errors or similar. This is primarly for showing how it works.