Refactor Navigation* components
Since we want to check the functionalty and styling of the Navigation
components that were made for the summer test version, we need to fix
any issues we find. This is a larger refactoring of the code to pull the
NavigationItem out into it's own component, and make the styling work as
it should with this new component.