Fix text and icon color for the different states
Since the text and icon colors for the different states were wrong, they
need to be fixed. This updates them based on the states for 'selected',
'hover', 'pressed/click' and default/rest.

I considered doing this in JS, but after discussing with the other devs
decided the better and easier solution is to do it in CSS. In JS it
would also stop functioning if a user has JS turned off.