Fix/pxweb2-140: NavigationBar checkup (#266)
Since we want to check the functionality and styling of the Navigation
components that were made for the summer test version, we need to fix
any issues we find. This is a larger refactoring of the code to pull the
NavigationItem out into it's own component, and make the styling work as
it should with this new component.

Since the text and icon colors for the different states were wrong, they
need to be fixed. This updates them based on the states for 'selected',
'hover', 'pressed/click' and default/rest.

I considered doing this in JS, but after discussing with the other devs
decided the better and easier solution is to do it in CSS. In JS it
would also stop functioning if a user has JS turned off.