Patch/update dependencies (#268)
* Update dependencies within ranges and with Nx

Since we want to keep our dependencies up-to-date, we need to update
the dependencies regularly. This update the project dependencies within
the ranges in package.json, and with the nx migrate latest command. It
does not include indiviual packages.

This also has that bug where the npm cli either adds or removes the
"dev: true" on lots of packages. I have not found a solution, nor reason
why this happens. But it does not seem to do anything else.

* Remove blankspace to force chromatic build