fix/px-web2-361: Fix languages sv and nor (#306)
* Update and fix swedish translation file

* Regenerate arabic translation file

The arabic translation file had alot of missing keys. I found a page
on the i18next site where you could use google translate to auto-gen a
new locale file in the targeted language.

Again, this is all from google translate, and have to be redone before
any actual use. Only here for our testing purposes for RTL-languages.

* Update and fix Nor language file

* Use locales for Information in ContentTop

* Expand the Modal to take custom keys for buttons

Since the Norwegian translations for the button texts in the Modal do
not match the meaning of the generic button tags, we need to expand the
Modal to support custom translation tags.

* Update select placeholder translations

* Update i18n-interface types