Add API call and state cleaning for new codelists
This adds an API call that returns the new values for a chosen codelist
in the VariableBox, and plugs that into the existing code for showing
new codelists. This means the VariableBox and state now can handle
chaning codelists and rendering them. However, the sync to the table is
not yet finished. It now stops showing any new values chosen, and shows
very little.

It also resets the chosen codelists when changing a language, since we
cannot reuse values for codelists between languages.