Add logic for changing codelists
This makes a new API-call and changes the values of a variable with a
selected codelist, when the user makes a new codelist selection.

The functinality for resetting to the default selection when switching
languages have not been implemented yet.

This code is missing the fix for the updated API, to get the correct
metadata. So it will still show the incorrect data on first load.