Remove som interal state in VariableBoxSelect
Since the variant of the Select component used in the variablebox is
it's own component, we do not need the same internal state management as
the default Select component. It will only get in the way of how we want
to update that componentent from the VaraiblesProvider state in the
Selection component/page.