Update the handleCodeChange to handle errors
Since we need to be able to handle whenever the api call fails when
changing codelists, we need to add the logic for this. This handles that
logic, but does not handle showing anything to the user when it happens.
That is something we need to speak with the designer about how it should
be handled.