Add basic fading when making api-call for new codelist
Since we need to indicate to the users when we are waiting on the api
response, when they change a codelist. We need some basic functionality
for this while the designer looks into how we want this to function in
the end. This is a temporary fix, which is similar to the table fading.