Feature/pxweb2 198 add ability to change codelist on variable (#336)
* Add API call and state cleaning for new codelists

This adds an API call that returns the new values for a chosen codelist
in the VariableBox, and plugs that into the existing code for showing
new codelists. This means the VariableBox and state now can handle
chaning codelists and rendering them. However, the sync to the table is
not yet finished. It now stops showing any new values chosen, and shows
very little.

It also resets the chosen codelists when changing a language, since we
cannot reuse values for codelists between languages.

* WIP: Removed changes for codelist defaultSelection

The task has been divided, and the part of getting and showing the
defaultSelection for the CodeLists have been moved to a different task.
This cleans up the branch and commits the current WIP changes for the
new scope of the task.

* Add logic for changing codelists

This makes a new API-call and changes the values of a variable with a
selected codelist, when the user makes a new codelist selection.

The functinality for resetting to the default selection when switching
languages have not been implemented yet.

This code is missing the fix for the updated API, to get the correct
metadata. So it will still show the incorrect data on first load.

* Tiny tmp fix

* Remove unneeded reset

* Remove reduntant check

* Revert new types for selection mapper

* Add basic reset of selection when changing language

* Add comments for reset on language change

* Remove old comment

* Remove som interal state in VariableBoxSelect

Since the variant of the Select component used in the variablebox is
it's own component, we do not need the same internal state management as
the default Select component. It will only get in the way of how we want
to update that componentent from the VaraiblesProvider state in the
Selection component/page.

* Bugfix: Make the function no longer update the original array

* Update the handleCodeChange to handle errors

Since we need to be able to handle whenever the api call fails when
changing codelists, we need to add the logic for this. This handles that
logic, but does not handle showing anything to the user when it happens.
That is something we need to speak with the designer about how it should
be handled.

* Fix a bug due to race condition with virtuoso

There was a problem with a race condition when rendering kodelist
values. This adds a check that fixes this, though we may want to look
closer at this later. Now atleast the app does not crash.

* Update the previous fix

Since Virtuoso "heavily" complains if we return null or similar, this
modifies the previous fix to return an empty div. This div is not
rendered, but instead removed by virtuoso before hand.

* Linting changes

* Add basic fading when making api-call for new codelist

Since we need to indicate to the users when we are waiting on the api
response, when they change a codelist. We need some basic functionality
for this while the designer looks into how we want this to function in
the end. This is a temporary fix, which is similar to the table fading.

* Left from merge from main

* Remove unused style code in selection.tsx

* Add resetting of variablebox when changing codelist

This adds resetting some states in the variableboxes when you change
codelists in the variable. This fixes some issues with the virtuoso list
and search remembering states into a new codelist.

* Add new storybook location to gitignore