Update to latest testing-library in pxweb2-ui
When we removed Nx we ran into a problem where tests were failing on a
clean install, and this had to do with the new major version of
testing-library/react. Running the older version worked. Tried to update
to the newest version now, and it both runs the tests and builds fine.